+++ src/offspot_config/builder.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import re
 from pathlib import PurePath as Path
 from typing import Any
-from urllib.parse import urlsplit
 
 from offspot_config.catalog import app_catalog, get_app_path
 from offspot_config.constants import CONTENT_TARGET_PATH
@@ -14,7 +13,6 @@ from offspot_config.inputs.str import BlockStr
 from offspot_config.oci_images import OCIImage
 from offspot_config.packages import AppPackage, FilesPackage, ZimPackage
 from offspot_config.utils.dashboard import Link, Reader
-from offspot_config.utils.download import read_checksum_from
 from offspot_config.utils.sizes import (
     get_margin_for,
     get_min_image_size_for,
@@ -209,23 +207,12 @@ class ConfigBuilder:
 
         # Add files for requested readers
         for reader in self.dashboard_readers:
-            checksum = None
-            # download.kiwix.org is known to provide digests via mirrorbrain
-            if urlsplit(reader.download_url).netloc == "download.kiwix.org":
-                try:
-                    checksum = Checksum(
-                        algo="md5",
-                        value=read_checksum_from(f"{reader.download_url}.md5"),
-                    )
-                # we cant assume this this work forever
-                except Exception:
-                    ...
             self.add_file(
                 url_or_content=reader.download_url,
                 to=str(KIWIXSERVE_DATA_PATH / reader.filename),
                 via="direct",
                 size=reader.size,
-                checksum=checksum,
+                checksum=reader.checksum,
                 is_url=True,
             )