+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- [inputs.checksum] `Checksum` gets a `to_dict()` method
+- [dashboard] `Reader` gets an optional `checksum`
+
+### Changed
+
+- [builder] Removed useless download.kiwix.org special behavior for reader checksum. Checksum now on Reader
+
 ## [2.0.0] - 2024-04-16
 
 ### Added